ARROW-9652: Move CREATE EXTERNAL TABLE planning to planner.rs
